Tidy up tests in preparation for adding more
